# .well-known/openapi.yaml
openapi: 3.0.3
info:
  title: Google Search Plugin API
  description: A Flask app that uses the Google Custom Search API to search the web.
  version: 1.0.0

servers:
  - url: http://localhost:5000
paths:
  /search:
    get:
      operationId: searchGet
      summary: Search Google and fetch HTML content
      description: |
        Searches Google using the provided query and returns the search results, along with the inner text of the links.
        The first 3 search results are returned in the `full_content` property, while the next 7 search results only contain the `summary` property.
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
          example: python
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                          description: The title of the search result
                        link:
                          type: string
                          format: uri
                          description: The URL of the search result
                        summary:
                          type: string
                          description: A summary of the content of the search result (available for the fourth through tenth results)
                        full_content:
                          type: string
                          description: The entire text content of the search result (available for the first three results)
                example:
                  results:
                    - title: Welcome to Python.org
                      link: https://www.python.org/
                      summary: The official home of the Python Programming Language...
                      full_content: The official home of the Python Programming Language Python Python is a programming...
                    - title: Python (programming language) - Wikipedia
                      link: https://en.wikipedia.org/wiki/Python_(programming_language)
                      summary: Python is an interpreted, high-level, general-purpose programming language...
                      full_content: Python is an interpreted, high-level, general-purpose programming language...
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                example:
                  error: No query provided
        'default':
          description: Error fetching search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                example:
                  error: Error fetching search results
  /search-deep:
    get:
      operationId: searchDeepGet
      summary: Search Google and fetch summaries of the first 3 pages of results based on the given query
      description: |
        Searches Google using the provided query and returns the search results, along with the inner text summary of the links.
        All results are returned in the `summary` property. Summaries are based on the query given.
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
          example: python
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                          description: The title of the search result
                        link:
                          type: string
                          format: uri
                          description: The URL of the search result
                        summary:
                          type: string
                          description: A summary of the content of the search result (available for all results)
                        full_content:
                          type: string
                          description: The entire text content of the search result (will be None for all deep results)
                example:
                  results:
                    - title: Welcome to Python.org
                      link: https://www.python.org/
                      summary: The official home of the Python Programming Language
                      full_content: None
                    - title: Python (programming language) - Wikipedia
                      link: https://en.wikipedia.org/wiki/Python_(programming_language)
                      summary: Python is an interpreted, high-level, general-purpose programming language
                      full_content: None
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                example:
                  error: No query provided
        'default':
          description: Error fetching search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                example:
                  error: Error fetching search results